requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django~=5.0.6
django-filter==24.3
django-oauth-toolkit==3.0.1
djangorestframework~=3.15.1
drf-spectacular~=0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
oauthlib==3.2.2
psycopg2-binary==2.9.10
pycparser==2.22
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
django-cors-headers
drf-yasg~=1.21.7